Reorder pn532.cpp Move mifare classic functions
